backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-decouple==3.8
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.9.0.post0
httptools==0.6.0
python-multipart==0.0.6
requests==2.31.0
email-validator==2.1.0
textblob==0.18.0
nltk==3.9.1
pymongo==4.6.1
